+++ mw/msg.ts
@@ -63,7 +63,7 @@ export const msg = (key: string, ...args: string[]): string => mw.msg(`cm-mw-${k
  * 为所有链接添加`target="_blank"`
  * @param $dom 容器
  */
-const blankTarget = ($dom: JQuery<HTMLElement>): JQuery<HTMLElement> => {
+const blankTarget = ($dom: JQuery): JQuery => {
 	$dom.find('a').add($dom.filter('a')).attr('target', '_blank');
 	return $dom;
 };
@@ -73,9 +73,9 @@ const blankTarget = ($dom: JQuery<HTMLElement>): JQuery<HTMLElement> => {
  * @param key 消息键，省略`cm-mw-`前缀
  * @param text 是否输出为文本
  */
-function parseMsg(key: string): JQuery<HTMLElement>;
+function parseMsg(key: string): JQuery;
 function parseMsg(key: string, text: true): string;
-function parseMsg(key: string, text?: true): string | JQuery<HTMLElement> {
+function parseMsg(key: string, text?: true): string | JQuery {
 	const message = mw.message(`cm-mw-${key}`);
 	return text ? message.parse() : blankTarget(message.parseDom());
 }
@@ -86,7 +86,7 @@ export {parseMsg};
  * @param key 消息键，省略`cm-mw-`前缀
  * @param args 替换`$1`等的参数
  */
-const notify = async (key: string, ...args: string[]): Promise<JQuery<HTMLElement>> => {
+const notify = async (key: string, ...args: string[]): Promise<JQuery> => {
 	const $p = blankTarget($('<p>', {html: msg(key, ...args)}));
 	await mw.notify($p, {type: 'success', autoHideSeconds: 'long'});
 	return $p;
@@ -98,7 +98,7 @@ const notify = async (key: string, ...args: string[]): Promise<JQuery<HTMLElemen
  * @param addons 新插件
  */
 export const welcome = async (baseVersion: string, addons: string[]): Promise<void> => {
-	let notification: JQuery<HTMLElement> | undefined;
+	let notification: JQuery | undefined;
 	if (!version) { // 首次安装
 		notification = await notify('welcome');
 	} else if (addons.length > 0) { // 更新版本